pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "churnops"
version = "0.1.0"
description = "Production-style churn prediction baseline for the ChurnOps MLOps platform."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115,<1.0",
    "joblib>=1.3,<2.0",
    "mlflow>=2.16,<4.0",
    "pandas>=2.1,<3.0",
    "prometheus-client>=0.21,<1.0",
    "PyYAML>=6.0,<7.0",
    "scikit-learn>=1.4,<2.0",
    "uvicorn>=0.32,<1.0",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.10,<3.0",
]
dev = [
    "build>=1.2,<2.0",
    "httpx>=0.27,<1.0",
    "pytest>=8.0,<9.0",
    "ruff>=0.6,<1.0",
    "wheel>=0.43,<1.0",
]

[project.scripts]
churnops-serve = "churnops.api.app:main"
churnops-train = "churnops.pipeline.train:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]
